+++ gfx/resources/OtInstances.h
@@ -45,7 +45,7 @@ public:
 	void add(const glm::mat4& instance, bool updateVersion=true);
 
 	// access individual instances
-	glm::mat4& operator[](int i) { return instances->operator[](i); }
+	glm::mat4& operator[](size_t i) { return instances->operator[](i); }
 	size_t size() { return instances->size(); }
 
 	// submit instances to GPU
